fix ModelAsModule.create_and_register_submodel for existing module/classes
The issue here is that specific model types have specific setup
code, and we must make sure that this code gets called even when
#new_submodel is not called or, in the case of ClassAsModel, when
a new class is created.

This splits the model object creation from the setup code. The setup
code is passed to base_model#setup_submodel(submodel), which replaces
the ill-defined and not-used existing #setup_model method.