define #clear_model
The existing strategy for clearing models was to deregister all
known submodels of a root model. However, once we introduced the
concept of permanent models, it became a bit contrived and quite
complicated to understand what should be removed and what should not.

In Roby, #clear_model actually wants to completely deregister
all non-permanent AND user-defined models. The API it needs is
actually at the level of the single model ("clear that model") instead
of at the level of the hierarchy.

Keep the old API for now, but add a (very simple) #clear_model