fix model de-registration w.r.t. the new #clear_model method
In principle, #clear_model can be called in any order. This has
the effect that a parent model might be cleared before its
children.

The current scheme was checking in #deregister_submodels if
there was indeed something deregistered, and was doing nothing
if it was not. This does not work as model Y could not have
the submodels registered anymore, but Y's supermodel still does.

Don't try to be smart anymore, and do all operations regardless
of everything else