add DSLs.find_through_method_missing
This is a helper for syskit-style access to find_X methods
through method missing, namely calling

  find_port("name")

when

  name_port

is called.

This helper properly takes care of all the error cases,
making sure that all these handlers are consistent with
each other, and avoiding the pitfall of infinite recursions
common in method_missing implementations.