dsls: add a cleaner implementation of find_through_method_missing
The current implementation of #find_through_method_missing was doing
a lot of computations which could often be made statically. Given that
it ends on the hot path, this is less than ideal.

Moreover, it led to deep call chains if one was calling it at each level
in method_missing, or required some prior knowledge about where the
call would be made.

The new implementation is done at the class level. The find_* calls
are declared and MetaRuby's find call is implemented in a module that
gets included in the target class. This ensures that a given class would
be able to filter valid matches quickly (one regexp match and one hash
access), and that the find_through_method_missing resolution logic needs
to be injected only once in the whole method_missing chain.