feat: implement the 'cache' argument to inherited_attribute
When enabled, the update model moves from a 'pull' model, where the attributes
are fetched from the ancestry, to a 'push' model, where the system automatically
updates the submodels. This improves performance in the principal use-case of
metaruby (modelling, where we change little but read a lot). In particular it
turns existence checks ('has_*') into a direct check at the submodel level.